requirements.txt
Django==4.2.3
requests
djangorestframework==3.14.0
django-rest-knox==4.2.0
django-cors-headers==4.2.0
django-filter==23.2
django-celery-beat==2.5.0
celery==5.3.1
redis>=3.5
pytest-cov==4.1.0
pytest-django==4.5.2
mixer==7.2.2